requirements.txt
wikipedia
googlesearch-python
Wikipedia-API
